v1.0.2 cu login cu compareTO + gresit->default + adaugare in baza de date meciuri+ v.2.0+ lista meciuri jucate + lista meciuri care se joaca+ cote de tip double

+++ src/main/java/com/example/demo/Domain/CotaGame.java
@@ -22,8 +22,8 @@ public class CotaGame {
     private Date data;
     private String team1;
     private String team2;
-    private int cotaTeam1;
-    private int cotaTeam2;
+    private double cotaTeam1;
+    private double cotaTeam2;
     private String turneu;
     private String tara;
